occlusion_masking: renaming method names
